+++ src/hash_equals.php
@@ -10,24 +10,7 @@
  */
 
 if (!function_exists('hash_equals')) {
-    
-    /**
-     * A strlen() implementation that won't get killed by mbstring.func_overload
-     * 
-     * @param @string $str
-     * @return int
-     */
-    function hash_compat_strlen($str)
-    {
-        static $exists = null;
-        if ($exists === null) {
-            $exists = function_exists('mb_strlen');
-        }
-        if ($exists) {
-            return mb_strlen($str, '8bit');
-        }
-        return strlen($str);
-    }
+    defined('USE_MB_STRING') or define('USE_MB_STRING', function_exists('mb_strlen'));
 
     /**
      * hash_equals — Timing attack safe string comparison
@@ -66,8 +49,17 @@ if (!function_exists('hash_equals')) {
             return false;
         }
 
+        // Ensures raw binary string length returned
+        $strlen = function($string) {
+            if (USE_MB_STRING) {
+                return mb_strlen($string, '8bit');
+            }
+
+            return strlen($string);
+        };
+
         // Compare string lengths
-        if (($length = hash_compat_strlen($known_string)) !== hash_compat_strlen($user_string)) {
+        if (($length = $strlen($known_string)) !== $strlen($user_string)) {
             return false;
         }